+++ internal/services/advisor/advisor_recommendations_data_source.go
@@ -10,7 +10,6 @@ import (
 	"github.com/hashicorp/go-azure-helpers/resourcemanager/commonschema"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/clients"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/suppress"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/validation"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/timeouts"
 )
@@ -36,7 +35,6 @@ func dataSourceAdvisorRecommendations() *pluginsdk.Resource {
 						string(advisor.Cost),
 						string(advisor.OperationalExcellence),
 					}, false),
-					DiffSuppressFunc: suppress.CaseDifferenceV2Only,
 				},
 			},