+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 ## 3.7.0 (Unreleased)
 
+FEATURES:
+
+* **New Resource**: `azurerm_elastic_cloud_elasticsearch` [GH-14821]
+* **New Data Source**: `azurerm_elastic_cloud_elasticsearch` [GH-14821]
+
 ENHANCEMENTS:
 
+* `azurerm_linux_function_app` - Add support for PowerShell 7.2  [GH-16718]
 * `azurerm_spring_cloud_app` - support for the `identity` block [GH-16280]
+* `azurerm_windows_function_app` - Add support for PowerShell 7.2  [GH-16718]
+
+
+BUG FIXES:
+
+* Data Source: `azurerm_databricks_workspace` - Handle panic when SKU field is missing [GH-16819]
+* `azurerm_express_route_gateway` - Handle gateway connections not found error [GH-16804]
 
 ## 3.6.0 (May 12, 2022)